requirements.txt
selenium>=4.26.0
fastapi>=0.116.0
uvicorn>=0.35.0
google-api-python-client>=2.174.0
google-auth>=2.40.0
python-dotenv>=1.1.1
